Move cluster and asyncssh module under ucp.benchmarks
